[Feat] #9 - 추천 콜렉션뷰 분리
